requirements.txt
pandas
python-dotenv
flask
flask-sqlalchemy
flask-login
flask-migrate
dash
dash-core-components
dash-html-components
dash-table